+++ agora_node_ext/agora_video_source.cpp
@@ -420,7 +420,6 @@ namespace agora{
         {
             if (m_initialized && m_peerJoined){
                 CaptureScreenByDisplayCmd cmd;
-
                 cmd.screenId = screenId;
                 cmd.regionRect = regionRect;
                 cmd.captureParams = captureParams;