+++ ts/Api/index.ts
@@ -40,7 +40,12 @@ import {
   VideoEncoderConfiguration,
   UserInfo,
   RendererOptions,
-  Metadata
+  Metadata,
+  AREA_CODE,
+  STREAM_PUBLISH_STATE,
+  STREAM_SUBSCRIBE_STATE,
+  RTMP_STREAMING_EVENT,
+  EncryptionConfig
 } from './native_type';
 import { EventEmitter } from 'events';
 import { deprecate, config, Config } from '../Utils';
@@ -697,6 +702,34 @@ class AgoraRtcEngine extends EventEmitter {
       fire('rtmpStreamingStateChanged', url, state, errCode);
     })
 
+    this.rtcEngine.onEvent('firstLocalAudioFramePublished', function(elapsed: number) {
+      fire('firstLocalAudioFramePublished', elapsed);
+    })
+
+    this.rtcEngine.onEvent('firstLocalVideoFramePublished', function(elapsed: number) {
+      fire('firstLocalAudioFramePublished', elapsed);
+    })
+  
+    this.rtcEngine.onEvent('audioPublishStateChanged', function(channel: string, oldState: STREAM_PUBLISH_STATE, newState: STREAM_PUBLISH_STATE, elapseSinceLastState: number) {
+      fire('audioPublishStateChanged', channel, oldState, newState, elapseSinceLastState)
+    })
+
+    this.rtcEngine.onEvent('videoPublishStateChanged', function(channel: string, oldState: STREAM_PUBLISH_STATE, newState: STREAM_PUBLISH_STATE, elapseSinceLastState: number) {
+      fire('videoPublishStateChanged', channel, oldState, newState, elapseSinceLastState)
+    })
+  
+    this.rtcEngine.onEvent('audioSubscribeStateChanged', function(channel: string, oldState: STREAM_PUBLISH_STATE, newState: STREAM_PUBLISH_STATE, elapseSinceLastState: number) {
+      fire('audioSubscribeStateChanged', channel, oldState, newState, elapseSinceLastState)
+    })
+  
+    this.rtcEngine.onEvent('videoSubscribeStateChanged', function(channel: string, uid: number, oldState: STREAM_SUBSCRIBE_STATE, newState: STREAM_SUBSCRIBE_STATE, elapseSinceLastState: number) {
+      fire('videoSubscribeStateChanged', channel, uid, oldState, newState, elapseSinceLastState)
+    })
+ 
+    this.rtcEngine.onEvent('rtmpStreamingEvent', function(url: string, eventCode: RTMP_STREAMING_EVENT) {
+      fire('rtmpStreamingEvent', url, eventCode);
+    })
+    
     this.rtcEngine.registerDeliverFrame(function(infos: any) {
       self.onRegisterDeliverFrame(infos);
     });
@@ -977,8 +1010,8 @@ class AgoraRtcEngine extends EventEmitter {
    *  - `ERR_INVALID_APP_ID (101)`: The app ID is invalid. Check if it is in 
    * the correct format.
    */
-  initialize(appid: string): number {
-    return this.rtcEngine.initialize(appid);
+  initialize(appid: string, areaCode: AREA_CODE = 0xFFFFFFFF): number {
+    return this.rtcEngine.initialize(appid, areaCode);
   }
 
   /**
@@ -4863,7 +4896,15 @@ class AgoraRtcEngine extends EventEmitter {
 
   setMaxMetadataSize(size: number): number {
     return this.rtcEngine.setMaxMetadataSize(size);
-  }  
+  }
+
+  sendCustomReportMessage(id: string, category: string, event: string, label: string, value: number): number {
+    return this.rtcEngine.sendCustomReportMessage(id, category, event, label, value);
+  }
+
+  enableEncryption(encryptionEnable: boolean, config: EncryptionConfig) : number {
+    return this.rtcEngine.enableEncryption(encryptionEnable, config);
+  }
 }
 /** The AgoraRtcEngine interface. */
 declare interface AgoraRtcEngine {
@@ -5875,11 +5916,39 @@ declare interface AgoraRtcEngine {
   ) => void): this;
   on(evt: 'receiveMetadata', cb: (
     metadata: Metadata
-    ) => void): this;
+  ) => void): this;
 
   on(evt: 'sendMetadataSuccess', cb: (
     metadata: Metadata
-    ) => void): this;
+  ) => void): this;
+
+  on(evt: 'firstLocalAudioFramePublished', cb: (
+    elapsed: number
+  ) => void): this;
+
+  on(evt: 'firstLocalVideoFramePublished', cb: (
+    elapsed: number
+  ) => void): this;
+
+  on(evt: 'audioPublishStateChanged', cb: (
+    channel: string, oldState: STREAM_PUBLISH_STATE, newState: STREAM_PUBLISH_STATE, elapseSinceLastState: number
+  ) => void): this;
+
+  on(evt: 'videoPublishStateChanged', cb: (
+    channel: string, oldState: STREAM_PUBLISH_STATE, newState: STREAM_PUBLISH_STATE, elapseSinceLastState: number
+  ) => void): this;
+
+  on(evt: 'audioSubscribeStateChanged', cb: (
+    channel: string, uid: number, oldState: STREAM_SUBSCRIBE_STATE, newState: STREAM_SUBSCRIBE_STATE, elapseSinceLastState: number
+  ) => void): this;
+
+  on(evt: 'videoSubscribeStateChanged', cb: (
+    channel: string, uid: number, oldState: STREAM_SUBSCRIBE_STATE, newState: STREAM_SUBSCRIBE_STATE, elapseSinceLastState: number
+  ) => void): this;
+
+  on(evt: 'rtmpStreamingEvent', cb: (
+    url: string, eventCode: RTMP_STREAMING_EVENT
+  ) => void): this;
 
   on(evt: string, listener: Function): this;
 }
@@ -6138,7 +6207,44 @@ class AgoraRtcChannel extends EventEmitter
     ) => {
         fire('connectionStateChanged', state, reason);
     });
-    
+
+    this.rtcChannel.onEvent('audioPubishStateChange', (
+      oldState: STREAM_PUBLISH_STATE,
+      newState: STREAM_PUBLISH_STATE,
+      elapseSinceLastState: number
+    ) => {
+      fire('audioPubishStateChange', oldState, newState, elapseSinceLastState)
+    })
+
+    this.rtcChannel.onEvent('videoPubishStateChange', (
+      oldState: STREAM_PUBLISH_STATE,
+      newState: STREAM_PUBLISH_STATE,
+      elapseSinceLastState: number
+    ) => {
+      fire('videoPubishStateChange', oldState, newState, elapseSinceLastState)
+    })
+
+    this.rtcChannel.onEvent('audioSubscribeStateChange', (
+      uid: number,
+      oldState: STREAM_SUBSCRIBE_STATE,
+      newState: STREAM_SUBSCRIBE_STATE,
+      elapseSinceLastState: number
+    ) => {
+      fire('audioSubscribeStateChange', uid, oldState, newState ,elapseSinceLastState)
+    })
+
+    this.rtcChannel.onEvent('videoSubscribeStateChange', (
+      uid: number,
+      oldState: STREAM_SUBSCRIBE_STATE,
+      newState: STREAM_SUBSCRIBE_STATE,
+      elapseSinceLastState: number
+    ) => {
+      fire('videoSubscribeStateChange', uid, oldState, newState ,elapseSinceLastState)
+    })
+
+    this.rtcChannel.onEvent('rtmpStreamingEvent', function(url: string, eventCode: RTMP_STREAMING_EVENT) {
+      fire('rtmpStreamingEvent', url, eventCode);
+    })
   }
   /**
    * Joins the channel with a user ID.
@@ -7519,6 +7625,37 @@ declare interface AgoraRtcChannel {
   on(evt: 'connectionStateChanged', cb: (
     state: ConnectionState,
     reason: ConnectionChangeReason
+  ) => void): this;
+
+  on(evt: 'audioPubishStateChange', cb: (
+    oldState: STREAM_PUBLISH_STATE,
+    newState: STREAM_PUBLISH_STATE,
+    elapseSinceLastState: number
+  ) => void): this;
+
+  on(evt: 'videoPubishStateChange', cb: (
+    oldState: STREAM_PUBLISH_STATE,
+    newState: STREAM_PUBLISH_STATE,
+    elapseSinceLastState: number
+  ) => void): this;
+
+  on(evt: 'audioSubscribeStateChange', cb: (
+    uid: number,
+    oldState: STREAM_SUBSCRIBE_STATE,
+    newState: STREAM_SUBSCRIBE_STATE,
+    elapseSinceLastState: number
+  ) => void): this;
+
+  on(evt: 'videoSubscribeStateChange', cb: (
+    uid: number,
+    oldState: STREAM_SUBSCRIBE_STATE,
+    newState: STREAM_SUBSCRIBE_STATE,
+    elapseSinceLastState: number
+  ) => void): this;
+
+  on(evt: 'rtmpStreamingEvent', cb: (
+    url: string, 
+    eventCode: RTMP_STREAMING_EVENT
   ) => void): this;
 }